scheduler and worker modules created and implementation examples are given in example folder
