+++ JobQue/example/scheduler/index.js
@@ -1,0 +1,9 @@
+const app = require('config')
+const cxnOptions = app.get('rethinkdb')
+var qOptions = app.get('qOptions')
+
+const scheduler = require('../../scheduler/scheduler')
+const main_scheduler = new scheduler({
+  cxnOptions: cxnOptions,
+  qOptions: qOptions
+})
