+++ JobQue/example/scheduler/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "example_scheduler",
+  "version": "1.0.0",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "bhumilsarvaiya",
+  "license": "ISC",
+  "dependencies": {
+    "chokidar": "^1.7.0",
+    "config": "^1.26.2",
+    "fs": "0.0.1-security",
+    "lodash": "^4.17.4",
+    "pino": "^4.7.2",
+    "rethinkdb": "^2.3.3",
+    "rethinkdb-job-queue": "^3.1.4"
+  },
+  "devDependencies": {
+    "webpack": "^3.5.6"
+  },
+  "description": ""
+}
