+++ JobQue/example/worker/mylog
@@ -1,0 +1,1 @@
+{"pid":10237,"hostname":"OBLAP-71","level":30,"time":1509441445643,"msg":"process in flow instance updated","v":1}
