+++ JobQue/example/worker/select_worker.js
@@ -3,7 +3,12 @@ const cxnOptions = app.get('rethinkdb')
 var qOptions = app.get('jOptions')
 qOptions.name = 'select_worker'
 
-const worker = require('./worker')(cxnOptions, qOptions, (input) => {
+const worker = require('../../worker/worker')
+const select_worker = new worker({
+    cxnOptions: cxnOptions,
+    qOptions: qOptions
+  },
+  (input) => {
 
   let output = {}
   output.rejected = []
