+++ JobQue/example/worker/task_worker.js
@@ -3,4 +3,8 @@ const cxnOptions = app.get('rethinkdb')
 var qOptions = app.get('jOptions')
 qOptions.name = 'task_worker'
 
-const worker = require('./worker')(cxnOptions, qOptions)
+const worker = require('../../worker/worker')
+const task_worker = new worker({
+  cxnOptions: cxnOptions,
+  qOptions: qOptions
+})
