+++ JobQue/scheduler/config.json
@@ -1,0 +1,46 @@
+{
+  "rethinkdb": {
+    "host": "localhost",
+    "port": 28015,
+    "db": "workers"
+  },
+  "qOptions": {
+    "name": "scheduler",
+    "masterInterval": 60000,
+    "changeFeed": true,
+    "concurrency": 1,
+    "removeFinishedJobs": false
+  },
+  "flowz_table": "flowzinstance",
+  "scheduler_table": "scheduler",
+  "system_logs_table": "system_logs",
+  "qJobTimeout": 3600000,
+  "qJobRetryMax": 0,
+  "chokidar": {
+    "persistent": true,
+    "ignoreInitial": false,
+    "followSymlinks": true,
+    "cwd": ".",
+    "disableGlobbing": false,
+    "usePolling": true,
+    "interval": 100,
+    "binaryInterval": 300,
+    "alwaysStat": false,
+    "depth": 99,
+    "awaitWriteFinish": {
+      "stabilityThreshold": 2000,
+      "pollInterval": 100
+    },
+    "ignorePermissionErrors": false,
+    "atomic": true
+  },
+  "pinoDB": {
+    "timestamp": "pino.stdTimeFunctions.slowTime"
+  },
+  "pinoConsole": {
+    "timestamp": false,
+    "prettyPrint": {
+      "forceColor": true
+    }
+  }
+}
