+++ JobQue/scheduler/package.json
@@ -1,7 +1,6 @@
 {
-  "name": "generic_worker",
+  "name": "flowz-scheduler",
   "version": "1.0.0",
-  "description": "",
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
@@ -9,9 +8,7 @@
   "author": "bhumilsarvaiya",
   "license": "ISC",
   "dependencies": {
-    "asyncawait": "^1.0.6",
     "chokidar": "^1.7.0",
-    "config": "^1.26.2",
     "fs": "0.0.1-security",
     "lodash": "^4.17.4",
     "pino": "^4.7.2",
@@ -20,5 +17,6 @@
   },
   "devDependencies": {
     "webpack": "^3.5.6"
-  }
+  },
+  "description": ""
 }
