--- JobQue/worker/mylog
@@ -1,1 +1,0 @@
-{"pid":4384,"hostname":"OBLAP22","level":40,"time":1509085244831,"msg":"all inputs not available","fId":"a3c56987-3c3f-49c4-a066-23ede63731b3","jobId":"StartEvent_1","v":1}
